Introduce `_math_utils`
